fix ClientSessionImpl@waitFor timeout much faster than configured
On some hosts, on which waiting for the `futureLock.wait(remWait)` is faster than 1 millisecond, `remWait` gets reduced by 123 milliseconds although not even one passed.
This causes a `ClientSessionEvent.TIMEOUT` after a much smaller time passed than the configured timeout.
As each host has different cycle times, this happens an unknown amount of times.
With these changes, the `remWait` is calculated based on the milliseconds passed instead of transformed nanoseconds.