Remove generics from key decoders
These classes attempted to statically enforce matching private and
public key classes via generics, but this never worked correctly. In
particular it fails for method parameters, and as result we had to
use unchecked conversions or raw types, and had to suppress warnings
because of that.

Moreover, this forces key classes (or interfaces) into the generic
parameters of subclasses, which gives trouble with key types that don't
exist on all Java versions supported (like ed25519).

So remove these generics and just use PrivateKey and PublicKey. We can
do so since we also removed any operations to get a decoder based on a
concrete key class -- this is bad practice anyway, and using either the
key algorithm or the SSH key type to figure out what kind of key a
particular instance is.