Simplify ed25519 handling
A lot of ed25519 operations can actually be done in a way that does not
depend on the concrete implementation (Bouncy Castle or net.i2p or even
Java built-in). Public keys are encoded as X.509 and private keys as
PKCS#8, and both encodings do contain the raw key bytes. Parsing these
formats and extracting the key bytes is simple, so we can write keys to
files or buffers easily. We can also reconstruct keys from the raw bytes
by creating valid X.509 or PKCS#8 KeySpecs.

The only minor hiccup is that the net.i2p provider uses non-standard
algorithm names for keys and signatures. So refactor this to generally
use the standard names, and let the SecurityProviderRegistrar create the
SecurityEntityFactory. The EdDSASecurityProviderRegistrar can then
create factories for net.i2p that replace the standard names by the
net.i2p-specific algorithm names.