More ed25519 simplifications
Introduce a PublicKeyFactory interface for something that can recover a
public key from a private key. On the ScurityProviderRegistrars
implement this interface, and use that in SecurityUtils to recover
ed25519 public keys.

With that, the EdDSASupport interface and its per-registrar instances
become obsolete. Remove them altogether.

Move all EdDSA things that are not specific to a provider (which is
actually everything but public key recovery) into the "eddsa.generic"
package.