GH-585: add a PublicKeyFactory for JCE ed25519 private keys
With this commit, sshd-common becomes a multirelease JAR, and so does
sshd-osgi.

On Java 15+, we can actually use the JDK's own ed25519 implementation,
so we also need a way to recover a public key given only a private key.

This recovery operation is actually only mandatory for reading an
ed25519 key pair from a PKCS#8 PEM private key. For reading key pairs
from an OpenSSH private key file, it's optional, since the OpenSSH
format contains both the private and the public key.

The MR-JAR setup requires dropping the source and target compiler
options. Use only and exclusively "release" now. (We can do so since
we require Java 17 for building anyway.)