Add SunECSecurityProviderRegistrar
Disabled by default it can be used to force the use of JCE algorithms
even if Bouncy Castle is present. Users are responsible themselves for
enabling consistent sets of algorithms. Mix-and-match between SunEC and
BC will not work (for instance EC keys from SunEC with EC signatures
from BC).

By default the registrar has ed25519 enabled, if supported by the Java
version it's running on. So if this registrar is enabled and Java has
ed25519, it will be used instead of Bouncy Castle ed25519.