Force SunEC for ed25519 public key recovery on Java15+
The hack feeding the private key bytes into KeyPairGenerator works only
with SunEC. It may not work with other security providers.

If SunEC should not be available, ed25519 public key recovery will not
be supported. This should have an effect at most on reading PEM files
that do not contain the public key. (Which they should, though,
according to RFC 5915: "Though the ASN.1 indicates publicKey is
OPTIONAL, implementations that conform to this document SHOULD always
include the publicKey field.")

[1] https://datatracker.ietf.org/doc/html/rfc5915#section-3