sshd-test: fix multiple executions with testcontainers
testcontainers removes test container images asynchronously when the
test JVM shuts down. Subsequent executions may then get into a situation
where they want to re-use an image that just gets removed, and then fail
to start the container.

Prevent that by ensuring that the first layer in each of these
containers is different for each JVM run. This is a stop-gap measure;
it has the disadvantage that we'll re-build the container for each
execution. Maybe a better approach is to not let testcontainers reap the
containers except in the last execution. (This would be possible through
ImageFromDockerfile(String, boolean), but we'd have to give the
containers static names, and somehow figure out how to set the flag to
true in the last run, which could become complicated if we want it to
figure out which profiles are active at all.)

Interestingly this problem only cropped up with failsafe, but not with
multiple executions of surefire.