SftpTransferTest: log exceptions
I got some flaky runs on this test. No idea what happened, the server
first gets a "connection reset" IOException, and then later logs a
"IllegalStateException: Not initialized", unfortunately without stack
trace.

The additional logging will include the stack trace, which will be
contained in the test XML for the flaky run. Hopefully that gives some
insight as to where this IllegalStateException is coming from.