GH-502: Do not load/create BouncyCastleSecurityProvider reflectively
It's not necessary since we have the optional dependency anyway. So
standard classloading is good enough.

For FIPS support, we still have to load the BouncyCastleFipsProvider
reflectively. Get the class via Class.forName(), though. This loads
the class through the normal mechanism that is also used for Java
imports, so if we can load it that way, we have access to the BC
classes we need. If we got through the thread context classloader, we
might be able to load it, but later still fail to resolve imports.

If a provider is registered already in the system make sure that we then
can access the matching provider class. If both FIPS and non-FIPS BC
providers are accessible (shouldn't happen), default to the FIPS one.

If FIPS mode is explicitly set consider only the FIPS provider.