GH-803: use JDK built-in ML-KEM on Java24+
Add KEMs to SecurityUtils and SecurityEntityFactories. Configure the
"SunJCEWrapper" SecurityProviderRegistrar to include support for ML-KEM
if it is available in Java's SunJCE security provider.

Rename interface KeyEncapsulationMethod to KEM: this serves now as a
wrapper around different KEM implementations from Bouncy Castle or from
the JDK. Because the class name is user-visible in the security provider
registrar configuration, using KEM minimizes confusions.

Move the Bouncy Castle implementations from sshd-core to sshd-common
where the rest of BC-specific stuff lives. Add a JceKEM factory to
create KEM instances if supported by the platform. This default factory
always throws NoSuchAlgorithmException.

In the multi-release section for Java 24, add a real JceKEM factory that
returns instances of our own KEM wrapper class that internally use
javax.crypto.KEM.

Change the Github workflows to use Java 25 for compiling, and run the
tests on Java 8, 17, and 25. (Previously we tested on 8, 11, and 17.)

The new JDK ML-KEM support is tested in the "multirelease" execution of
the maven-failsafe-plugin in bundle sshd-test in test OpenSshMlKemTest.