GH-830: Get EC curve parameters from the security provider
For EC public keys read from a ByteArrayBuffer, Bouncy Castle would
generate X.509 encodings using as algorithm parameters not the named
curve OID but the raw curve parameters.

This was caused by our using manually constructed ECParameterSpecs with
the values given in [1]. SunEC has code matching raw parameters back to
known curve names, and then using the OID, but Bouncy Castle does not.

So instead of constructing the ECParameterSpecs explicitly let the
security provider provide them. With that approach, Bouncy Castle can
produce a spec that also transports the OID, and then uses the OID in
the X.509 encoding.